Implements the hit and death sounds
Player has a crash sound for losing their first two lives and an explosion sound for losing their last life. The enemies make an alien screech and explode. It's pretty quiet but it might need to be more quiet, im not sure rn. The autoloaded sounds might need to be expanded to the player and boss as well when we actually make scenes for destroying them.