Implements more sounds, game over screens, and boss death anim
There is now a sound effect that plays when you hit an enemy and it does not die. There are now game screens for game over and victory. Both display final score. There is music for both. The boss death sequence has been updated so that it flies around while exploding and screeching before the victory screen appears.